Reset password is working
If user forgets password and inputs the wrong password, they get a link to the reset password page, they then need to submit their email. They will get a link in their email to reset the password, which will take them to a page where they can reset it.

Also had to make a minor change the routing for reset password in the app.py for